cmd/stringer: add support for flag -bitflag
When the new flag -bitflag is passed to the command,
stringer generates code for types treating their constants
as bit flags. The String methods generated return strings
of the form "(Flag0|Flag1)" when multiple bit flags are
set.

There is a -nocache flag to specify that the code generated
should not cache the computed strings. Generally the default
of using a cache makes the accesses significantly less costly
to runtime performance. The cache size is bounded to 256
entries. The cache size is not bounded by the length of the
computed strings.