cmd/stringer: bitflag code uses offsets
Switch from using indexes into the names string to using offsets.
The bitflag code generated iterates through the flag values from low to
high already anyway and the indexes often would not fit into a uint8 and
in theory might not fit in the next size either. As long as we enforce
that no one bitflag name may exceed 255 bytes, we can use uint8 offsets
instead. This doesn't change the code for the original constants, only
those specified as bitflags.